makeRentalTransaction method implemented

+++ src/models/Rentable.java
@@ -2,5 +2,5 @@ package models;
 
 public interface Rentable {
     
-    public double calculateRentalCost(int days);
+    public double calculateRentalCost(long days);
 }
